Merge conflicts resolved in:
* api/neteq/neteq.h
* audio/channel_receive.cc
* audio/channel_receive.h
* media/engine/simulcast_encoder_adapter.cc
* modules/audio_coding/BUILD.gn
* modules/audio_coding/acm2/acm_resampler.cc
* modules/audio_coding/acm2/acm_resampler_unittest.cc
* modules/audio_processing/audio_processing_impl.h
* modules/video_coding/BUILD.gn
* p2p/base/p2p_transport_channel_unittest.cc
* pc/jsep_transport_controller.cc
* pc/rtp_transport.h
* pc/rtp_transport_internal.h
* unsafe_buffers_paths.txt

+++ api/video/video_bitrate_allocation.cc
@@ -119,7 +119,7 @@ VideoBitrateAllocation::GetSimulcastAllocations() const {
     std::optional<VideoBitrateAllocation> layer_bitrate;
     if (IsSpatialLayerUsed(si)) {
       layer_bitrate = VideoBitrateAllocation();
-      for (int tl = 0; tl < kMaxTemporalStreams; ++tl) {
+      for (size_t tl = 0; tl < kMaxTemporalStreams; ++tl) {
         if (HasBitrate(si, tl))
           layer_bitrate->SetBitrate(0, tl, GetBitrate(si, tl));
       }
